### Challenge
This is an analysis for a Kickstart Campaign analyzing launch date, backers and seasonality for the most optimum launch of the project.  All data is sourced from attached Excel workbook and the analysis is summarized in the attached Powerpoint Presentation.  I have included the .pnp slides for your review, although they are included in the workbook.